Time: 99 ms (99.06%), Space: 64.5 MB (58.59%) - LeetHub

+++ 0997-find-the-town-judge/0997-find-the-town-judge.cpp
@@ -7,7 +7,7 @@ public:
     }
     
     int findJudge(int n, vector<vector<int>>& trust) {
-        int trustCount[1001] = {0};
+        short trustCount[1001] = {0};
         int size = trust.size();
         for(int i=0;i<size;++i){
             ++trustCount[trust[i][1]];
